Move NavBar code to AppBar component
